style: fix lint

+++ src/interfaces/compound/CEtherI.sol
@@ -8,10 +8,13 @@ interface CEtherI is CTokenI {
 
     function redeem(uint256 redeemTokens) external returns (uint256);
 
-    function liquidateBorrow(address borrower, CTokenI cTokenCollateral) external payable;
+    function liquidateBorrow(address borrower, CTokenI cTokenCollateral)
+        external
+        payable;
 
-    function borrow(uint256 borrowAmount) external returns (uint);
+    function borrow(uint256 borrowAmount) external returns (uint256);
 
     function mint() external payable;
+
     function repayBorrow() external payable;
 }
